Refactor filesystem handling and update imports for clarity
