--- esp_data/__init__.py
@@ -1,3 +1,0 @@
-# from esp_data.io.paths import AnyPath, GSPath, R2Path
-
-# __all__ = ["AnyPath", "GSPath", "R2Path"]
