+++ esp_data/io/__init__.py
@@ -1,3 +1,4 @@
+from esp_data.io.parsers import read_audio
 from esp_data.io.paths import AnyPath, GSPath, R2Path
 
-__all__ = ["AnyPath", "GSPath", "R2Path"]
+__all__ = ["AnyPath", "GSPath", "R2Path", "read_audio"]
