+++ tests/test_paths.py
@@ -2,7 +2,8 @@ from pathlib import PosixPath
 
 import pytest
 
-from esp_data.io import AnyPath, is_cloud_path, is_local_path
+from esp_data.io import AnyPath
+from esp_data.io.paths import is_cloud_path, is_local_path
 
 
 def test_local_path():
